[#112] Replace qualifier in manifests by build timestamp
When the pom.xml files are changed to use a milestone qualifier as part
of their version, the MANIFEST.MF and feature.xml files have also be
adjusted to contain a qualified version. It is not desired to have the
milestone qualifier (e.g. 2.11.1.M1) in the p2 artifacts, but have a
build time qualifier here. In contrast, the artifacts deployed to OSSRH
should have the milestone qualifier, so this must be the POM version.

Change-Id: If3ea2b140752492f6b68467b443c51cee2bb3f23
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>