Cleanup actions for old nightlies and milestones
When building a release old milestone update sites are deleted and the
composite site descriptor for the milestones is re-created. The last 3
milestones are kept.

On every build old nightly and milestone drops are deleted from the
download area. Here it must be considered that directories are located
in a nested structure and that sorting by paths is not sufficient. This
would make a '2.9' version appear younger than a '2.1x' version.

Change-Id: I1b7b0fe4cf1fdfd8ec6267d555e522db97ab9900
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>