[Setup] Rework target-platform definition and project-import
Define MWE's target by only using the 'org.eclipse.emf.mwe2.target'
target-file as composed-target in the targlet-task. This allows to
remove the out-dated and not (anymore) working list of repositories as
well as all the variables only used in these lists.
Furthermore this allows to remove the explicit listing of TP content,
like Guava or other eclipse features.

The project-import through the targlet-task is removed and the explicit
project import is moved before the targlet-task in order to make the
target-file referenced in the latter available in the workspace.